--- tests/Promise/Fixtures/EntityWithConstants.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Cycle\ORM\Promise\Tests\Fixtures;
-
-class EntityWithConstants
-{
-    const MY_CONST = 'value';
-}
