--- tests/Promise/Fixtures/EntityWithoutConstants.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Cycle\ORM\Promise\Tests\Fixtures;
-
-class EntityWithoutConstants
-{
-}
