+++ tests/Promise/ProxyPrinter/Fixtures/EntityWithTrait.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace Cycle\ORM\Promise\Tests\ProxyPrinter\Fixtures;
+
+class EntityWithTrait
+{
+    use TestTrait;
+}
