+++ tests/Promise/ProxyPrinter/Fixtures/EntityWithoutConstructor.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Cycle\ORM\Promise\Tests\Fixtures;
+namespace Cycle\ORM\Promise\Tests\ProxyPrinter\Fixtures;
 
 class EntityWithoutConstructor
 {
