+++ tests/Promise/ProxyPrinter/Fixtures/EntityWithoutTrait.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Cycle\ORM\Promise\Tests\ProxyPrinter\Fixtures;
+
+class EntityWithoutTrait
+{
+}
